+++ packages/cloud-cognitive/src/components/PageHeader/PageHeader.js
@@ -317,7 +317,7 @@ export let PageHeader = React.forwardRef(
     );
 
     useWindowResize(() => {
-      // on window resieze and other updates some values may have changed
+      // on window resize and other updates some values may have changed
       checkUpdateVerticalSpace();
     }, [
       actionBarItems,
@@ -423,7 +423,7 @@ export let PageHeader = React.forwardRef(
           ])}
           ref={headerRef}>
           <Grid>
-            <div className={`${blockClass}__non-navigation-row-contnet`}>
+            <div className={`${blockClass}__non-navigation-row-content`}>
               {hasBreadcrumbRow ? (
                 <Row
                   className={cx(`${blockClass}__breadcrumb-row`, {
@@ -810,16 +810,7 @@ PageHeader.propTypes = {
       PropTypes.shape({
         ...Button.propTypes,
         key: PropTypes.string.isRequired,
-        kind: PropTypes.oneOf([
-          'primary',
-          'secondary',
-          'danger',
-          'ghost',
-          'danger--primary',
-          'danger--ghost',
-          'danger--tertiary',
-          'tertiary',
-        ]),
+        kind: Button.propTypes.kind,
         label: PropTypes.node,
         onClick: PropTypes.func,
       })